- All the downloadmanager changes from the win32 ver. - become Warning Free(tm) again -- mainly just casting args to sprintf()

+++ io/src/pipeline.cpp
@@ -181,5 +181,5 @@ void PipelineUnit::DebugPrint(void)
     printf("[%c %c %ld]", 
         m_bSleeping ? 'S' : 'W',
         m_bPause ? 'P' : 'R',
-        m_pOutputBuffer->GetBufferPercentage());
+        (long int)m_pOutputBuffer->GetBufferPercentage());
 }
